WindowServer: Cache the last two backing stores to make flipping faster.
Instead of reattaching to the shared buffer on every flip, keep a reference
to the last seen backing store GraphicsBitmap. This way we can simply
swap the two last buffers when a flip happens.

This does increase memory consumption in apps that disable double-buffering
but I'll address that issue separately.